requirements.txt
# Web Framework
Flask==3.1.2

# Deep Learning
torch==2.1.0
torchvision==0.16.0
torchmetrics==1.2.0

# Segmentation Models
segmentation-models-pytorch==0.3.3

# Image Processing
opencv-python==4.8.1.78
Pillow==10.1.0
albumentations==1.3.1

# Scientific Computing
numpy==1.24.3
matplotlib==3.8.0

# LLM Integration (Optional - for XAI explanations)
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# Additional Utilities
python-dotenv==1.0.0